replace content of create_schema.py by that of create_schema_2.py, deleted create_schema_2.py, improved comments of create_uischema.py
